requirements.txt
Flask
requests
numpy
Pillow
matplotlib
scikit-learn
tensorflow

